push down to zinc compile
