+++ src/python/pants/backend/jvm/tasks/jvm_compile/zinc/zinc_compile.py
@@ -16,6 +16,8 @@ from pants.backend.jvm.subsystems.java import Java
 from pants.backend.jvm.subsystems.jvm_platform import JvmPlatform
 from pants.backend.jvm.subsystems.scala_platform import ScalaPlatform
 from pants.backend.jvm.subsystems.zinc import Zinc
+from pants.backend.jvm.targets.jar_library import JarLibrary
+from pants.backend.jvm.targets.jvm_app import JvmApp
 from pants.backend.jvm.targets.jvm_target import JvmTarget
 from pants.backend.jvm.targets.scalac_plugin import ScalacPlugin
 from pants.backend.jvm.tasks.classpath_entry import ClasspathEntry
@@ -25,6 +27,7 @@ from pants.base.build_environment import get_buildroot
 from pants.base.exceptions import TaskError
 from pants.base.hash_utils import hash_file
 from pants.base.workunit import WorkUnitLabel
+from pants.build_graph.resources import Resources
 from pants.engine.fs import (
     EMPTY_DIRECTORY_DIGEST,
     DirectoryToMaterialize,
@@ -324,11 +327,11 @@ class BaseZincCompile(JvmCompile):
         if isinstance(target, ScalacPlugin):
             result[_SCALAC_PLUGIN_INFO_FILE] = textwrap.dedent(
                 """
-          <plugin>
-            <name>{}</name>
-            <classname>{}</classname>
-          </plugin>
-        """.format(
+                <plugin>
+                  <name>{}</name>
+                  <classname>{}</classname>
+                </plugin>
+                """.format(
                     target.plugin, target.classname
                 )
             )
@@ -935,3 +938,36 @@ class ZincCompile(BaseZincCompile):
 
     def select_source(self, source_file_path):
         return source_file_path.endswith(".java") or source_file_path.endswith(".scala")
+
+    def calculate_jvm_modulizable_targets(self):
+
+        if not self.context.products.is_required_data("export_dep_as_jar_signal"):
+            return set()
+
+        def is_jvm_or_resource_target(t):
+            return isinstance(t, (JvmTarget, JvmApp, JarLibrary, Resources))
+
+        jvm_and_resources_target_roots = set(
+            filter(is_jvm_or_resource_target, self.context.target_roots)
+        )
+        jvm_and_resources_target_roots_minus_synthetic_addresses = set(
+            t.address for t in filter(lambda x: not x.is_synthetic, jvm_and_resources_target_roots)
+        )
+        all_targets = set(self.context.targets())
+        modulizable_targets = set(
+            t
+            for t in self.context.build_graph.transitive_dependees_of_addresses(
+                jvm_and_resources_target_roots_minus_synthetic_addresses,
+                predicate=lambda x: x in all_targets,
+            )
+            if is_jvm_or_resource_target(t)
+        )
+        synthetic_modulizable_targets = set(filter(lambda x: x.is_synthetic, modulizable_targets))
+        if len(synthetic_modulizable_targets) > 0:
+            raise TaskError(
+                f"Modulizable targets must not contain synthetic target, but in this case {synthetic_modulizable_targets}.\n"
+                f"It means that certain thrift target(s) depends back onto the targets you want to import to IDE."
+            )
+
+        self.context.products.get_data("jvm_modulizable_targets", set).update(modulizable_targets)
+        return modulizable_targets
